# [WIP] Box complex types in AnyValue enum