@lalitb 2024-07-08
AnyValue benchmark results:

| Benchmark                        | PR Time (ns) | Main Time (ns) |
|----------------------------------|--------------|----------------|
| CreateOTelValueStaticString      | 3.7731       | 3.7642         |
| CreateOTelAnyValueStaticString   | 4.4106       | 34.740         |
| CreateOTelValueString            | 15.083       | 15.393         |
| CreateOTelAnyValueString         | 16.076       | 41.257         |
| CreateOTelValueInt               | 3.1472       | 3.1382         |
| CreateOTelAnyValueInt            | 4.1100       | 33.097         |