@lalitb 2024-07-09
And the [LogRecord](https://github.com/open-telemetry/opentelemetry-rust/blob/main/opentelemetry-sdk/benches/log.rs) benchmark. The benchmark for box'ed complex types has gone up slightly as expected (eg for `simple-log-with-bytes/no-context` , `simple-log-with-map-any-value`), there is  visible improvement for simple types, though not to the scale of `AnyValue` benchmark alone.


| Benchmark                               | PR Branch Time (ns) | Main Branch Time (ns) |
|-----------------------------------------|---------------------|-----------------------|
| simple-log/no-context                   | 123.56              | 141.61                |
| simple-log/with-context                 | 123.84              | 140.62                |
| simple-log-with-int/no-context          | 142.48              | 156.29                |
| simple-log-with-int/with-context        | 143.36              | 155.16                |
| simple-log-with-double/no-context       | 140.69              | 155.43                |
| simple-log-with-double/with-context     | 147.05              | 153.54                |
| simple-log-with-string/no-context       | 142.76              | 156.71                |
| simple-log-with-string/with-context     | 143.79              | 154.11                |
| simple-log-with-bool/no-context         | 144.85              | 161.52                |
| simple-log-with-bool/with-context       | 144.18              | 160.25                |
| simple-log-with-bytes/no-context        | 165.61              | 171.69                |
| simple-log-with-bytes/with-context      | 172.59              | 173.45                |
| simple-log-with-a-lot-of-bytes/no-context| 168.16             | 175.58                |
| simple-log-with-a-lot-of-bytes/with-context| 170.43           | 174.53                |
| simple-log-with-vec-any-value/no-context| 204.47              | 219.25                |
| simple-log-with-vec-any-value/with-context| 207.93            | 218.56                |
| simple-log-with-inner-vec-any-value/no-context| 273.05         | 313.19                |
| simple-log-with-inner-vec-any-value/with-context| 273.08       | 287.55                |
| simple-log-with-map-any-value/no-context| 217.71              | 223.04                |
| simple-log-with-map-any-value/with-context| 219.62            | 223.06                |
| simple-log-with-inner-map-any-value/no-context| 300.22         | 298.58                |
| simple-log-with-inner-map-any-value/with-context| 301.72       | 300.53                |
| long-log/no-context                     | 124.05              | 141.42                |
| long-log/with-context                   | 123.79              | 140.95                |
| full-log/no-context                     | 95.383              | 116.95                |
| full-log/with-context                   | 111.75              | 125.99                |
| full-log-with-4-attributes/no-context   | 200.20              | 179.25                |
| full-log-with-4-attributes/with-context | 202.43              | 176.71                |
| full-log-with-9-attributes/no-context   | 338.03              | 305.05                |
| full-log-with-9-attributes/with-context | 337.06              | 280.23                |
| full-log-with-attributes/no-context     | 430.26              | 419.97                |
| full-log-with-attributes/with-context   | 428.65              | 413.89                |