@lalitb 2024-07-09
Updating benchmark with LogRecord creation for 4 attributes:
| Benchmark                                      | Main Repo (ns)                   | PR (box'ed) (ns)              | PR (box'ed + smallVec) (ns)     |
|------------------------------------------------|----------------------------------|-------------------------------|---------------------------------|
| full-log-with-4-attributes/no-context          | 165.94 - 166.51                  | 159.96 - 160.40               | 110.21 - 110.36                 |
| full-log-with-4-attributes/with-context        | 186.79 - 187.43                  | 176.67 - 176.97               | 133.45 - 133.91                 |