@utpilla 2024-07-09
> Updating benchmark with LogRecord creation for 4 attributes:
> 
> Benchmark	Main Repo (ns)	PR (box'ed) (ns)	PR (box'ed + smallVec) (ns)
> full-log-with-4-attributes/no-context	165.94 - 166.51	159.96 - 160.40	110.21 - 110.36
> full-log-with-4-attributes/with-context	186.79 - 187.43	176.67 - 176.97	133.45 - 133.91

@lalitb I think we need to `black_box` the static string values provided in the `log_record.add_attribute` method to see the desired perf difference.

```rust
log_benchmark_group(c, "full-log-with-4-attributes", |logger| {
        let mut log_record = logger.create_log_record();
        ...
        log_record.add_attribute("name", black_box("my-event-name"));
        log_record.add_attribute("event.id", black_box(20));
        log_record.add_attribute("user.name", black_box("otel"));
        log_record.add_attribute("user.email", black_box("otel@opentelemetry.io"));
        logger.emit(log_record);
    });
```

I think the huge perf difference that shows up for simply creating `AnyValue` for static strings showed up because you wrapped the `AnyValue` creating with a `black_box`. I believe if you don't make any changes to the enum and run the create values from static string benchmarks without `black_box`, you would get comparable results already.

```rust
c.bench_function("CreateOTelAnyValueStaticStr", |b| {
        b.iter(|| {
            let _v = black_box(AnyValue::String("value1".into()));
        });
    });
```